// apps/web/src/types.ts
export type Fillup = Tables<'fillups'>;
// ...
export type FillupDTO = Pick<
  Fillup,
  | 'id'
  | 'car_id'
  | 'date'
  | 'fuel_amount'
  | 'total_price'
  | 'odometer'
  | 'distance_traveled'
  | 'fuel_consumption'
  | 'price_per_liter'
>;
// ...
export interface PaginationDTO {
  next_cursor: string | null;
  has_more: boolean;
  total_count: number;
}